A bug with query interpretation of several queries fixed, generic loader rewritten

+++ lego-core/src/main/scala/ch/epfl/data/dblab/legobase/tpch/Queries.scala
@@ -350,7 +350,7 @@ object Queries {
         val hj4 = new HashJoinOp(soPartsupp, hj3)((x, y) => x.PS_PARTKEY == y.L_PARTKEY[Int] && x.PS_SUPPKEY == y.L_SUPPKEY[Int])(x => x.PS_PARTKEY)(x => x.L_PARTKEY[Int])
         val hj5 = new HashJoinOp(hj4, soOrders)((x, y) => x.L_ORDERKEY[Int] == y.O_ORDERKEY)(x => x.L_ORDERKEY[Int])(x => x.O_ORDERKEY)
         val aggOp = new AggOp(hj5, 1)(x => new Q9GRPRecord(x.N_NAME[LBString], dateToYear(x.O_ORDERDATE[Int])))(
-          (t, currAgg) => { currAgg + ((t.L_EXTENDEDPRICE[Double] * (1.0 - t.L_DISCOUNT[Double]))) - ((1.0 * t.PS_SUPPLYCOST[Double]) * t.L_QUANTITY[Int]) })
+          (t, currAgg) => { currAgg + ((t.L_EXTENDEDPRICE[Double] * (1.0 - t.L_DISCOUNT[Double]))) - ((1.0 * t.PS_SUPPLYCOST[Double]) * t.L_QUANTITY[Double]) })
         val sortOp = new SortOp(aggOp)((kv1, kv2) => {
           val k1 = kv1.key; val k2 = kv2.key
           val r = k1.NATION diff k2.NATION
@@ -624,11 +624,11 @@ object Queries {
         val scanPart = new SelectOp(new ScanOp(partTable))(x => x.P_CONTAINER === medbag && x.P_BRAND === brand15)
         val jo = new HashJoinOp(scanPart, scanLineitem)((x, y) => x.P_PARTKEY == y.L_PARTKEY)(x => x.P_PARTKEY)(x => x.L_PARTKEY)
         val wo = new WindowOp(jo)(x => x.L_PARTKEY[Int])(x => {
-          val sum = x.foldLeft(0.0)((cnt, e) => cnt + e.L_QUANTITY[Int])
+          val sum = x.foldLeft(0.0)((cnt, e) => cnt + e.L_QUANTITY[Double])
           val count = x.size
           val avg = 0.2 * (sum / count)
           x.foldLeft(0.0)((cnt, e) => {
-            if (e.L_QUANTITY[Int] < avg) cnt + e.L_EXTENDEDPRICE[Double]
+            if (e.L_QUANTITY[Double] < avg) cnt + e.L_EXTENDEDPRICE[Double]
             else cnt
           }) / 7.0
         })
@@ -722,11 +722,11 @@ object Queries {
         val jo = new SelectOp(new HashJoinOp(so2, so1)((x, y) => y.P_PARTKEY == x.L_PARTKEY)(x => x.L_PARTKEY)(x => x.P_PARTKEY))(
           x => x.P_BRAND[LBString] === Brand31 &&
             (x.P_CONTAINER[LBString] === SMBOX || x.P_CONTAINER[LBString] === SMCASE || x.P_CONTAINER[LBString] === SMPACK || x.P_CONTAINER[LBString] === SMPKG) &&
-            x.L_QUANTITY[Int] >= 4 && x.L_QUANTITY[Int] <= 14 && x.P_SIZE[Int] <= 5 || x.P_BRAND[LBString] === Brand43 &&
+            x.L_QUANTITY[Double] >= 4 && x.L_QUANTITY[Double] <= 14 && x.P_SIZE[Int] <= 5 || x.P_BRAND[LBString] === Brand43 &&
             (x.P_CONTAINER[LBString] === MEDBAG || x.P_CONTAINER[LBString] === MEDBOX || x.P_CONTAINER[LBString] === MEDPACK || x.P_CONTAINER[LBString] === MEDPKG) &&
-            x.L_QUANTITY[Int] >= 15 && x.L_QUANTITY[Int] <= 25 && x.P_SIZE[Int] <= 10 || x.P_BRAND[LBString] === Brand43 &&
+            x.L_QUANTITY[Double] >= 15 && x.L_QUANTITY[Double] <= 25 && x.P_SIZE[Int] <= 10 || x.P_BRAND[LBString] === Brand43 &&
             (x.P_CONTAINER[LBString] === LGBOX || x.P_CONTAINER[LBString] === LGCASE || x.P_CONTAINER[LBString] === LGPACK || x.P_CONTAINER[LBString] === LGPKG) &&
-            x.L_QUANTITY[Int] >= 26 && x.L_QUANTITY[Int] <= 36 && x.P_SIZE[Int] <= 15)
+            x.L_QUANTITY[Double] >= 26 && x.L_QUANTITY[Double] <= 36 && x.P_SIZE[Int] <= 15)
         val aggOp = new AggOp(jo, 1)(x => "Total")(
           (t, currAgg) => { currAgg + (t.L_EXTENDEDPRICE[Double] * (1.0 - t.L_DISCOUNT[Double])) })
         val po = new PrintOp(aggOp)(kv => { kv.key; printf("%.4f\n", kv.aggs(0)) }, () => true)
@@ -756,7 +756,7 @@ object Queries {
         val scanLineitem = new SelectOp(new ScanOp(lineitemTable))(x => x.L_SHIPDATE >= constantDate1 && x.L_SHIPDATE < constantDate2)
         val jo1 = new HashJoinOp(scanPart, scanPartsupp)((x, y) => x.P_PARTKEY == y.PS_PARTKEY)(x => x.P_PARTKEY)(x => x.PS_PARTKEY)
         val jo2 = new HashJoinOp(jo1, scanLineitem)((x, y) => x.PS_PARTKEY[Int] == y.L_PARTKEY && x.PS_SUPPKEY[Int] == y.L_SUPPKEY)(x => x.PS_PARTKEY[Int])(x => x.L_PARTKEY)
-        val aggOp = new AggOp(jo2, 1)(x => new Q20GRPRecord(x.PS_PARTKEY[Int], x.PS_SUPPKEY[Int], x.PS_AVAILQTY[Int]))((t, currAgg) => { currAgg + t.L_QUANTITY[Int] })
+        val aggOp = new AggOp(jo2, 1)(x => new Q20GRPRecord(x.PS_PARTKEY[Int], x.PS_SUPPKEY[Int], x.PS_AVAILQTY[Int]))((t, currAgg) => { currAgg + t.L_QUANTITY[Double] })
         val selOp = new SelectOp(aggOp)(x => { x.key.PS_PARTKEY; x.key.PS_SUPPKEY; x.key.PS_AVAILQTY > 0.5 * x.aggs(0) })
         val jo3 = new HashJoinOp(selOp, scanSupplier)((x, y) => x.key.PS_SUPPKEY == y.S_SUPPKEY)(x => x.key.PS_SUPPKEY)(x => x.S_SUPPKEY)
         val jo4 = new HashJoinOp(scanNation, jo3)((x, y) => x.N_NATIONKEY == y.S_NATIONKEY[Int])(x => x.N_NATIONKEY)(x => x.S_NATIONKEY[Int])
